Gson into file test moved to create file test
