+++ src/test/java/gson/CreateFileTest.java
@@ -1,5 +1,6 @@
 package gson;
 
+import com.google.gson.Gson;
 import org.junit.Test;
 
 import java.io.IOException;
@@ -16,6 +17,20 @@ public class CreateFileTest {
         String reader=createFile.readFile();
         assertEquals("input output data stream", reader );
 
+
+    }
+    @Test
+    public void gsonTest() throws IOException {
+        Gson gson =new Gson();
+        Employee employee=new Employee();
+        employee.setName("Ankit");
+        employee.setSalary(150);
+        employee.setTotalYearsInComp(2);
+        String employeeJson = gson.toJson(employee);
+        CreateFile createFile=new CreateFile();
+        createFile.writeFile(employeeJson);
+        String reader=createFile.readFile();
+        assertEquals("{\"name\":\"Ankit\",\"salary\":150,\"totalYearsInComp\":2}",reader);
 
     }
 }
